fix(sync_masternodes): improve error handling for masternode list updates
Changed best_height field type from i32 to u32 for consistency across the codebase:
- Updated PeerInfo struct in types.rs to use Option<u32> for best_height
- Modified mock.rs to cast best_height as u32 in peer_info()
- Updated peer.rs best_height() method to return Option<u32>
- Fixed connection.rs to store peer_best_height as Option<u32> and cast from i32 when needed

This ensures type consistency between peer_info() and get_peer_best_height() methods
and avoids potential issues with negative height values.

+++ dash-spv/src/network/connection.rs
@@ -39,7 +39,7 @@ pub struct TcpConnection {
     peer_version: Option<u32>,
     peer_services: Option<u64>,
     peer_user_agent: Option<String>,
-    peer_best_height: Option<i32>,
+    peer_best_height: Option<u32>,
     peer_relay: Option<bool>,
     peer_prefers_headers2: bool,
     peer_sent_sendheaders2: bool,
@@ -245,7 +245,7 @@ impl TcpConnection {
         self.peer_version = Some(version_msg.version);
         self.peer_services = Some(version_msg.services.as_u64());
         self.peer_user_agent = Some(version_msg.user_agent.clone());
-        self.peer_best_height = Some(version_msg.start_height);
+        self.peer_best_height = Some(version_msg.start_height as u32);
         self.peer_relay = Some(version_msg.relay);
         
         tracing::info!("Updated peer info for {}: height={}, version={}, services={:?}",
